Safari-611.1.10/LayoutTests/fast/box-shadow/hit-test-box-shadow-on-zero-width-clipping-container-expected.txt

PASS: target element is #container-with-width
